Fix laundering of Bad_power_state exn from Xenopsd
This exception, defined in Xenops_interface, only has two arguments: the
expected and actual power states. The code that transformed this into a XenAPI
error didn't have the VM's identifier to hand and therefore created an
Api_error with too few arguments. This led to errors like the following being
returned from the API:

  You attempted an operation on a VM that was not in an appropriate power state
  at the time; for example, you attempted to start a VM that was already
  running.  The parameters returned are the VM's handle, and the expected and
  actual VM state at the time of the call.
  vm: Halted
  expected: Running
  actual: <unknown>

This patch fixes this so that Xapi_xenops.transform_xenops_exn always knows the
ref of the VM related to the operation.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>