oasis: Start moving to Oasis build system
This gets us as far as building the code generator and running it to generate
the autogen code all driven by Oasis.

The settings used to generate the code are the ones used by default in the
relevant OMakefile although these were configurable so we might want to
maintain that.

Don't wnat to nuke the existing Makefile yet so to use this, move the existing
configure script and Makefile to one side and run:

    oasis setup && ./configure && make

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>