CP-17481: is_control_domain =/=> domain zero
The API has always had documentation that allows for more than one VM to
have `is_control_domain = true`:

> field ... "is_control_domain" "true if this is a control domain (domain 0 or a driver domain)";

Most of Xapi has been written with this in mind, e.g. `Xapi_host.destroy`
which iterates through a list, `my_control_domains`.

However there are a few places where it is assumed that dom0 is the only
VM with `is_control_domain = true` and this patch makes these a bit more
sane.

Where the original code assumed only one control domain and did an
operation that only makes sense for domain 0, an additional filter is
applied to only operate on VMs with `domid = 0`. Where the original code
assumed only one control domain and did an operation that makes sense
for all control domains then this has been extended to operate on the
list of VMs with `is_control_domain = true`.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>