CP-17481: Add upgrade logic for new Host.control_domain
This needs to be done before the "hi-level" database upgrade because the
code to ensure the VM record for dom0 exists is in the dbsync logic
which is called as part of starting the database engine. This all
happens before the DB upgrade, and therefore, the dbsync logic tries to
create an additional dom0 record because it does not have the
Host.control_domain link.

This patch adds the Host.control_domain link during the dbsync if it
finds a VM with the CONTROL_DOMAIN_UUID from the Xensource_inventory.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>